refactoring in core ui and image processing modules
