+++ core/core-ui/src/main/java/com/rempawl/core/ui/bottomsheet/ImageSourcePickerOption.kt
@@ -1,4 +1,4 @@
-package com.rempawl.bottomsheet
+package com.rempawl.core.ui.bottomsheet
 
 enum class ImageSourcePickerOption {
     CAMERA,
