+++ core/core-ui/src/main/java/com/rempawl/core/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.rempawl.theme
+package com.rempawl.core.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
