+++ data/image-processing/src/main/java/com/rempawl/image/processing/model/TextBlockWrapper.kt
@@ -1,9 +1,9 @@
-package com.rempawl.image.processing.usecase
+package com.rempawl.image.processing.model
 
 import android.graphics.Rect
 import com.google.mlkit.vision.text.Text
 
-data class TextBlockWrapper(val boundingBox: Rect? = null) {
+internal data class TextBlockWrapper(val boundingBox: Rect? = null) {
     companion object {
         fun from(textBlock: Text.TextBlock) = TextBlockWrapper(textBlock.boundingBox)
     }
