+++ data/image-processing/src/main/java/com/rempawl/image/processing/repository/MLKitDetectionRepositoryImpl.kt
@@ -7,7 +7,7 @@ import com.google.mlkit.vision.objects.DetectedObject
 import com.google.mlkit.vision.objects.ObjectDetector
 import com.google.mlkit.vision.text.TextRecognizer
 import com.rempawl.core.kotlin.extensions.EitherResult
-import com.rempawl.image.processing.usecase.TextBlockWrapper
+import com.rempawl.image.processing.model.TextBlockWrapper
 import kotlinx.coroutines.suspendCancellableCoroutine
 import kotlin.coroutines.resume
